libjlm: use assignment operation for updating the global state
The solution so far only used the symbol table to update global state. This
would not have worked in the presence of branches. The order in which
the nodes would have been traversed would have mattered.

Introduce an assignment statement that updates the global state and
therefore solves this problem.