jlm-rvsdg: fix bug in loop-restructuring
The restructuring did not work properly for CFGs where the exit
node does not post-dominate all basic blocks in the graph. Such
a graph can appear when there is a statically detectable endless
loop in the program. The front-end compiler might remove the exit
edge from such a loop, leaving us with an improperly structured
CFG from which we could not create an RVSDG. This fix provides a
solution to this problem by adding the edge back into the graph.