libjlm/ir: fix bug in branch reduction of aggregation
While reducing a branch, we removed all the inedges of the join
node. This is too much as the join node could also be the sink of other
edges. For example, the join could also be the head of a loop and have
an incoming back-edge. Such a back-edge was removed as well. This fix
ensures that only the inedges from the individual branch nodes are
removed.