libjlm/aa: fix bug in Steensgaard alias analysis
We got a segmentation fault when analyzing the pointer arguments of external calls. This was due to the usage of FindOrInsertRegisterLocation() in combination with Location::Cast<>(), which casted the returned location from FindOrInsertRegisterLocation() to a register location. However, it is not guaranteed that FindOrInsertRegisterLocation() returns a register location. This means that Location::Cast<> might return a nullptr, leading to the segmentation fault.