Create util library (#211)
Refactors the libjlm/util and libjive/util parts into their own library. This PR only moves the files around, adjust the headers, and moves the unit tests such that everything compiles again and the unit tests succeed. It does not adjust the namespaces. This will happen in a later PR.