Create rvsdg library (#212)
This PR renames the libjive library to librvsdg. It moves the files to the jlm/rvsdg subfolder, adjusts the header guards, adjusts all the include files, and move the unit tests to tests/jlm/rvsdg. It does not change any namespaces or other code. This will happen in a separate PR.