Move jlm-opt logic into JlmOptCommand class (#242)
The JlmOptCommand class represents the jlm-opt command line tool. This
PR moves the entire logic of the jlm-opt command line tool from the
jlm-opt.c file into the JlmOptCommand class. This permits for better
testing in the future and cleans up quite a few code smells.
Specifically, this PR makes the following changes:

1. It privatizes all attributes of the JlmOptCommandLineOptions class.
2. Establishes the JlmOptCommandLineOptions::OptimizationIds as the sole
truth for jlm-opt optimizations.
3. Ensures that the jlm-opt optimizations performed under the hood by
jlc are set in the command line parser and not later in the pipeline.
4. Moves the jlm-opt logic into the JlmOptCommand class