Remove jlm-opt's single statistics file (#249)
The single statistics file could lead to problems when compiling in
parallel as multiple processes would write to the file. Instead we
create a unique statistics file per jlm-opt invocation. This PR does the
following:

1. Change the semantics of the -s command line option of jlm-opt from
specifying the statistics file to specifying a directory. This directory
is where the statistics file will end up in.
2. Ensure that unique statistics files are created per jlm-opt
invocation.