Fix possible C++ constructor ordering issue when creating test registry (#261)
The order of constructor invocation across translation units is not well
defined, though it seems fair to guess that the clang++ and g++ linkers,
and maybe even MSVC++ use the same order as the object files on the
command line. New incremental linkers and other tools may however not
expect those constraints.

This can cause the registration of tests to segfault, when the global
map is not ready yet. Putting it behind a function ensures that the
constructor is called before use.

Also fixes typos in the definition of the PHONY targets `check-ctets`
and `check-utets`.