Add overloads for rvsdg::view and PointsToGraph::ToDot to allow giving the same unique names to outputs in both (#297)
In practice this means the `PointsToGraph`'s dot output now adds an
extra part to `RegisterNode`s, as seen here:

![image](https://github.com/phate/jlm/assets/3748845/0335945b-9f11-47f4-b655-91ba0442a18b)

The names in parentheses correspond to the names given by rvsdg::view in
the following output:
```
[]{
  o0 := DELTA[MyList] 
    []{
      o1 := ConstantPointerNull 
    }[o1]
  o2 := LAMBDA[next] o0 
    [a3, a4, a5, a6 <= o0]{
      o7 := BITS32(0) 
      o8 := BITS32(4) 
      o9 o10 := ALLOCA[ptr] o8 
      o11 := MemStateMerge o10 a4 
      o12 o13 := LOAD a6 o11 
      o14 := STORE o9 o12 o13 
      o15 o16 := LOAD o9 o14 
      o17 := GetElementPtr o15 o7 o7 
      o18 o19 := LOAD o17 o16 
      o20 := STORE o9 o18 o19 
      o21 o22 := LOAD o9 o20 
    }[o21, a3, o22, a5]
}[o0, o2]
```

The example is generated by adding the following lines to the top of
`TestAndersen.cpp`'s `TestLinkedList()`:
```cpp
  std::unordered_map<const jlm::rvsdg::output*, std::string> outputMap;
  std::cout << jlm::rvsdg::view(test.graph().root(), outputMap) << std::endl;
  std::cout << jlm::llvm::aa::PointsToGraph::ToDot(*ptg, outputMap) << std::endl;
```

The functionality is added through overloads that allow specifying the
`unordered _map<const output*, std::string>` used to give outputs unique
names. The `ToDot` method has been modified to be able to take in such a
map, and use it while making output. It is not the cleanest approach
ever, but it was very helpful during debugging.

---------

Co-authored-by: Nico Reissmann <nico.reissmann@gmail.com>