Add function related constraints to PointerObjectSet + extra constraint logic (#299)
This PR comes on top of #298

Adds the two constraints `HandleEscapingFunctionConstraint` and
`FunctionCallConstraint`, that make use of the function `BiMap` to
communicate when a function can be called from outside the module, and
to communicate points-to-sets between callsite and callee.

Also adds the derived flags `CanPoint()` and `CanBePointee()` to the
`PointerObject`, since I found out that several memory objects shouldn't
be pointing to other memory objects, and it was best to make the logic
transparent.
When elements are attempted added to the points-to-set of something
where `CanPoint() == false`, it is a no-op, and silently ignored. This
allows the constraint application code to forget about them.