Add Andersen Alias Analysis pass (#300)
Builds on top of #298 and #299

Adds an Andersen pass, creating `PointerObject`s and constraints for a
given rvsdg-llvm-module, solving the constraints, and producing a
`PointsToGraph`.

All tests fro `TestSteensgaard.cpp` are ported over to
`TestAndersen.cpp`, except for the Statistics test. The solutions are of
course a bit different, mostly removing edges due to more precise
analysis, but also adding some, due to Steensgaard not registering
imports and lambdas as escaped / external.