Reduce runtime of points-to graph construction in Steensgaard alias analysis (#324)
This PR significantly reduces the runtime of the points-to graph
construction pass in the Steensgaard alias analysis.

In the old implementation, we created for each RVSDG pointer output a
register node in the points-to graph and then connected this node to the
respective points-to graph memory nodes. However, a lot of register
nodes point to the same subset of memory nodes as they are in the same
equivalence class. By creating individual register nodes in the
points-to graph for each output in the equivalence class and then
connecting each of these register nodes to the respective memory nodes,
we added a lot of redundant nodes and edges in the points-to graph. The
result was that points-to graph construction was quadratic and the
points-to graphs were huge in terms of nodes and edges.

This PR fixes this problem by adding RegisterSet nodes to the points-to
graph. These nodes can represent all pointer outputs that are in an
equivalence class. The result is that we are inserting only a single
points-to graph node for ALL registers in an equivalence class and only
connect this single node to the respective points-to graph memory nodes.
This significantly reduces the size of the points-to graphs and
points-to graph construction time in the Steensgaard analysis.

This PR makes the following contributions:
1. Introduces RegisterSet nodes to the points-to graph
2. Uses these nodes in the Steensgaard points-to graph construction
3. Refactors the Steensgaard points-to graph construction pass

The aim is to remove the old RegisterNode class from the points-to
graph, but for that we need to also remove their usage in the Andersen
alias analysis.

I did some experiments to see the indicative impact of this change. The
experiments were done with a jlm debug build. I only did a single run
for each file. Thus, take the runtime numbers with a grain of salt, but
they should be indicative enough.

The following graph shows the number of eliminated register nodes in the
points-to graph, i.e., NumRegisterNodesOldImplementation -
NumRegisterNodesNewImplementation:

![RegisterReductionAbsolute-AutoGenerated](https://github.com/phate/jlm/assets/153740/f5f2bb2b-ea9b-4d2d-a529-b98152923d49)

The following graph shows the percentage of eliminated register nodes in
the points-to graph between the old and new implementation, i.e., 100.0
- (100.0 * (NumRegisterNodesNewImplementation /
NumRegisterNodesOldImplementation)):

![RegisterReductionPercentage-AutoGenerated](https://github.com/phate/jlm/assets/153740/2e246d68-b732-4ed6-b83b-5cf3d3e9f254)

The following graph shows the runtime reduction for the Steensgaard
points-to graph construction pass in ms between the two implementations,
i.e., RuntimeOld - RuntimeNew:

![RuntimeAbsolute-AutoGenerated](https://github.com/phate/jlm/assets/153740/4e21d2d3-e2d3-4cd8-a0f7-e659c7fdab7b)

The following graph shows the runtime reduction for the Steensgaard
points-to graph construction in % between the two implementations, i.e.,
100.0 - (100.0 * (RuntimeNew / RuntimeOld)). Thus, a negative percentage
number means that the old implementation was faster, and a positive
percentage means that the new implementation is faster:

![RuntimePercentage-AutoGenerated](https://github.com/phate/jlm/assets/153740/2b6fba90-e3b8-4ff9-b680-7ebbf7886c55)