Make Andersen emit PointsToGraphs with RegisterSetNodes (#340)
Instead of emitting one `PointsToGraph::RegisterNode` per
`rvsdg::output*` represented in the PointsToGraph, it groups all
`rvsdg::output*`s associated with the same `PointerObject`, and creates
one `RegisterSetNode` for them. This should enable the future removal of
`RegisterNode`s.

While translating the tests, I observed that the number of
`RegisterSetNode`s was about 50-90% of the amount of `RegisterNode`s the
test's PointsToGraph used to have, a nice improvement.

---------

Co-authored-by: Nico Reissmann <nico.reissmann@gmail.com>