Propagate points-to flags between sets in Steensgaard (#350)
This PR introduces a points-to flag propagation pass in the Steensgaard
alias analysis such that the points-to flags are properly propagated in
the disjoint location set graph. So far, the points-to flags were only
annotated for each disjoint location set, but were not propagated to the
points-to set. However, all of the points-to flags are contagious, i.e.,
if a set is marked with a flag, then the points-to set must also be
marked with it.

This should put the points-to flags on solid footing as:
1. The points-to flags are added for each location on creation time.
2. The flags are propagated to the root element of a set when a location
is joined to a set.
3. The flags are propagated between sets with the just introduced
propagation pass.

This PR introduces the third point from above. The first and second
point are already present in the code.

Closes #335