Shell script for building CIRCT + CI update
The circt.sh script clones and builds to required commit of CIRCT.

The build happens in circt-build and the files are installed in
circt-build/circt.

The CI is updated to use this build script instead of depending on
the jlm-eval-suite repository.

The configure.sh script uses --enable-hls instead of --circt-path.