Fix lifetime of StructType::Declarations in test RVSDGs (#367)
The declarations of struct types in the Test RVSDGs only lived for as
long as we were in the SetupRvsdg() method. This is a problem as these
declarations also needed to be alive after this method. It lead to
dangling pointers. The only reason why valigrind-check did not catch
this was because we never accessed these pointers again by coincidence.

This PR does the following:
1. Adds the possibility of adding these declarations to the RVSDG and
InterProceduralGraph modules to extend their lifetime for as long as
they are needed in these modules. This also enables us to remove a FIXME
in the RVSDG construction, where these declarations were kept alive for
as long as jlm was running.
2. Fixes the lifetime problem in the test RVSDG.