Enable load-store reduction again (#369)
1. Fixes the load-store reduction. The problem was a missing check for
the stored and loaded value type being the same. If they are not the
same, then we might have to insert a truncation or conversion operation
instead of just bypassing the load and store. I marked the place in the
code with a FIXME such that we can improve upon it in the future.
2. Adds a unit test that exposes the problem.

Closes #302