Add measurement/timer system and sourceFile to Statistics base class (#387)
The Statistics part of #382. Modifies all existing Statistics to provide
the sourceFile path.

The mapping from id to name is now in `Statistics.cpp`, but many of the
existing statistics used names that were different from the C++ enums,
so I tried to make the classes keep their old names.

It might be better to have `GetName()` be an abstract method, and let
subclasses provide names that way, instead of a single BijectiveMap. Let
me know!