[Andersen] move Statistics to the new system, add PassConfig, rename Constraints (#395)
This comes on top of #393. Ignore those files for now.

`Andersen::Statistics` has been modified to use the new system. A lot of
the RVSDG-related labels are shared with Steensgaard, so I can move
those strings to `Labels` if desired.

The `PassConfiguration` is what allows the choice of multiple solver
implementation, and eventually specifying what kinds of offline
preprocessing should be done.

I renamed the two horribly named constraints to `LoadConstraint` and
`StoreConstraint`, since that is the only situation they are used, and
it aligns better with the literature.

---------

Co-authored-by: Nico Reissmann <nico.reissmann@gmail.com>