Add GraphWriter API for programatically building graphs and outputting as ASCII or dot (#403)
This is the API I ended up with, it uses some `friend` declarations to
ensure internal consistency, so the public API should be hard to misuse.

You create **graphs**, where you can add **nodes** and **edges**. These
can all have labels and arbitrary attributes set, which will be included
in the dot output. The ASCII output ignores attributes for brevity.

There are also **ArgumentNode**, **ResultNode** and **InOutNode** that
are nice when representing data flowing into the graph, through nodes,
and out of the graph. InOutNodes have an arbitrary number of input ports
and output ports, which can also have labels and attributes. They can
also have subgraphs.

The InOutNodes use the HTML table labels in dot. In dot, subgraphs are
rendered as separate graphs, so I was planning on customizing a
JavaScript library for GraphViz to make graphs inside nodes render
nicely when requested.

The ArgumentNodes and ResultNodes have the option of being connected to
ports in outside graphs. Otherwise, edges must go between ports within
the same graph.

All GraphElements have the option of being mapped to a program object
(any type of pointer), which serves two purposes: You can look up
elements by program object, to aid in building the graphs, and you can
refer to program objects by pointer in attributes, and have the pointer
be replaced by the short unique id of the graph element.

I'm very open to suggestions for changes to the interface, internals or
the output, as this should be something usable for everybody.

There is currently a helper function for setting the background color on
a node/port.
I could envision a lot more helper functions, to make it easier for the
end user to find useful and correct attributes, e.g.
```
edge.SetAttribute("style", "dashed");
// becomes
edge.SetStyle(EdgeStyle::Dashed);
```

The code takes care to escape strings in both regular dot and the HTML
sub-language.
I will of course add tests to all the functionality that it makes sense
to test.

Under follows some examples of API usage, and the corresponding output
in ASCII and rendered GraphViz:

<details>
<summary>A regular graph</summary>

```cpp
  jlm::util::GraphWriter writer;
  auto & graph = writer.CreateGraph();
  graph.SetAttribute("rankdir", "LR");

  auto & node = graph.CreateNode();
  node.SetLabel("DELTA");
  node.SetAttribute("shape", "oval");

  auto & node2 = graph.CreateNode();
  node2.SetLabel("ALLOCA");

  auto & edge = graph.CreateDirectedEdge(node, node2);
  edge.SetAttribute("style", "dashed");
  edge.SetAttribute("minlen", "3");
  edge.SetAttribute("headlabel", "Pointee");
  edge.SetAttribute("taillabel", "Pointer");

  auto & node3 = graph.CreateNode();
  node3.SetLabel("ALLOCA2");
  auto & edge2 = graph.CreateUndirectedEdge(node2, node3);
  edge2.SetLabel("Unified");
```
becomes
```
{
  node0:DELTA
  node1:ALLOCA<-[node0, node2]
  node2:ALLOCA2<-node1
}
```
and

![image](https://github.com/phate/jlm/assets/3748845/822ebfdc-7c4d-46a8-bc41-63a1756f6b40)
</details>

<details>
<summary>A dataflow graph</summary>

```cpp
  jlm::util::GraphWriter writer;
  auto & graph = writer.CreateGraph();
  auto & arg0 = graph.CreateArgumentNode();
  arg0.SetLabel("CTX");
  arg0.SetFillColor("yellow");
  auto & res0 = graph.CreateResultNode();

  auto & node0 = graph.CreateInOutNode(0, 1);
  node0.SetLabel("\"Cool Name!\"");
  node0.SetFillColor("gray");

  auto & node1 = graph.CreateInOutNode(3, 3);
  node1.SetLabel("<Main Node>");
  node1.GetOutputPort(1).SetFillColor("yellow");

  auto & node2 = graph.CreateInOutNode(2, 1);
  node2.SetLabel("End Node");

  graph.CreateDirectedEdge(arg0, node1.GetInputPort(2));

  graph.CreateDirectedEdge(node0.GetOutputPort(0), node1.GetInputPort(1));
  graph.CreateDirectedEdge(node0.GetOutputPort(0), node1.GetInputPort(2));

  graph.CreateDirectedEdge(node1.GetOutputPort(0), node2.GetInputPort(0));
  graph.CreateDirectedEdge(node1.GetOutputPort(1), node2.GetInputPort(1));

  graph.CreateDirectedEdge(node2.GetOutputPort(0), res0);
```
becomes
```
{
  ARG a0
  o0 := "\"Cool Name!\""
  o1, o2, o3 := "<Main Node>" [], o0, [a0, o0]
  o4 := "End Node" o1, o2
  RES o4
}
```
and

![image](https://github.com/phate/jlm/assets/3748845/bbeabe51-4be9-49cf-befb-a08557885d15)
</details>

<details>
<summary>A graph with a subgraph</summary>

```cpp
  jlm::util::GraphWriter writer;
  auto & graph = writer.CreateGraph();

  auto & source = graph.CreateNode();
  auto & node = graph.CreateInOutNode(3, 3);
  graph.CreateDirectedEdge(source, node.GetInputPort(0));
  graph.CreateDirectedEdge(node.GetOutputPort(0), node.GetInputPort(2));

  auto & subgraph = node.CreateSubgraph();
  auto & arg0 = subgraph.CreateArgumentNode();
  auto & arg1 = subgraph.CreateArgumentNode();
  arg0.SetOutsideSource(node.GetInputPort(0));
  arg1.SetOutsideSource(node.GetInputPort(2));

  auto & res0 = subgraph.CreateResultNode();
  auto & res1 = subgraph.CreateResultNode();
  res0.SetOutsideDestination(node.GetOutputPort(2));
  res1.SetOutsideDestination(node.GetOutputPort(1));

  subgraph.CreateDirectedEdge(arg0, res1);
  subgraph.CreateDirectedEdge(arg1, res0);
```
becomes
```
{
  node0:NODE
  o0, o1, o2 := NODE node0, [], o0
    {
      ARG a0 <= node0, a1 <= o0
      RES a1 => o2, a0 => o1
    }
}
```
and

![image](https://github.com/phate/jlm/assets/3748845/85d2e8c8-ad52-49f1-ad73-2e59d17ee3a6)
</details>

This PR is part of #308

---------

Co-authored-by: Nico Reissmann <nico.reissmann@gmail.com>