[Andersen] add PointerObjectConstraintSet solver using worklist algorithm (#419)
This PR comes on top of #418, I will move its base once that is merged.

This PR modifies the existing naive solver slightly to enable code
re-use using some templated functions.
It is also designed with future expansions in mind (e.g. difference
propagation and unification), which might explain some peculiar choices.

`HandleEscapingFunctionConstraint` is removed as a constraint, as it was
previously added for every function in the program. It made more sense
to make it a global constraint, and not something you individually add
per function.

The worklist solver is made the default in Andersen's configuration.
Currently, double-checking against the naive solver is controlled by an
environment variable, as it is purely a debug thing, and feels like a
nice mix between a compiler directive and a command line flag added to
jlm. Let me know if you prefer it in a different way.