Add TopDownMemoryNodeEliminator (#277)
This PR introduces the TopDownMemoryNodeEliminator. The
TopDownMemoryNodeEliminator restricts the number of regions memory
states need to be routed through by taking the lifetime of memory nodes
into account. For example, the lifetime of a memory node originating
from an `alloca` is at most the function the alloca is defined in.

The AgnosticMemoryNodeProvider and RegionAwareMemoryNodeProvider do not
take such liftime information into account when producing a memory node
provisioning. The idea of the TopDownMemoryNodeEliminator is to take the
memory node provisioning produced by these memory node providers as seed
provisioning and produce a new provisioning which takes the lifetime
information into account. This simple example illustrates the idea:

```
static void
f(int x)
{
  int* alloca;
  *alloca = x;
}
```

The AgnosticMemoryNodeProvider and the RegionAwareMemoryNodeProvider
would produce the following provisioning:

RegionEntry: f -> {allocaMemoryNode}
RegionExit: f -> {allocaMemoryNode}

Thus, they would consider the allocaMemoryNode alive at the region entry
even though the alloca node was not even defined yet.

The TopDownMemoryNodeEliminator would take this provisioning as input,
track the lifetime of alloca nodes, and restrict the provisioning to the
following:

RegionEntry: f -> {}
RegionExit: f- > {allocaMemoryNode}

Ultimately, this would mean that the MemoryStateEncoder would need to
route no state through the entry of function `f`.

Currently, the TopDownMemoryNodeProvider restricts the lifetime ONLY for
alloca nodes and ONLY for region entries.