CI HLS workflow uses specific jlm-eval-suite commit
The HLS workflow performs a checkout of a specific jlm-eval-suite
commit. This makes it possible to make updates to the
jlm-eval-suite independently of jlm, and that later can be used by
the jlm CI. This avoids breaking the CI until pull requests have
been approved for both the jlm-eval-suite and jlm repositories.