Ensure phi nodes containing delta nodes are properly handled in the ReginAwareMemoryNodeProvider (#440)
This PR fixes a bug where phi nodes containing delta nodes were not
properly handled in the RegionAwareMemoryNodeProvider. The PR contains
the following:

1. A creator for the ConstantStruct opertion to conveniently create it
in an RVSDG.
2. Makes AssignAndPropagateMemoryNodes a proper method instead of having
it as a lambda.
3. Adds an RVSDG test that exposes the problem.
4. Fixes the issue in the RegionAwareMemoryNodeProvider by simply not
adding delta nodes to the created region summaries. This is achieved by
simply moving code around, making the overall code also more intuitive
and easier to follow.

Closes #439