Add custom Hash functor for HashSet, with overload for std::pair (#452)
@phate's comment about using `std::set` instead of `util::HashSet` made
me remember why I picked `set` in the first place: `std::pair` doesn't
have a specialization of `std::hash`. Since we have our own `HashSet`
class that we like to use, I thought it would be fun to see what it
would look like to add support for `std::pair`without requiring the user
to specify a custom hash function for it.

I also added some constructors to be able to initialize the set with
sets that do not have the same hash functor.

If we prefer to not go this route that is of course completely OK. I do
enjoy being able to place `std::pair` in HashSets without any extra
hassle, though.