Enable load-load-state reduction again (#461)
I could not reproduce the problem from #307 any longer. Looking at the
input program and the reported error, I suspect that this optimization
was just the trigger, but not the actual cause. I believe that the cause
was the missing loaded type in the equality operator of the
LoadOperation. It was fixed in #456.

Closes #307