Introduce shared ptr for types (#499)
This change introduces shared_ptr for types, uses them in most places
for storage, and starts putting in singleton objects for several
primitive
types.

The singletons use the "untracked feature" for std::shared_ptr -- the
"object used for lifetime tracking" can be separate from the "pointer
that the shared_ptr logically points to". It is possible to make the
former a "null" object in which case we actually have a non-reference
counted "shared" object that is supposed to have eternal lifetime.

This small trick allows to skip the cost of refcount inc/dec for these
eternal primitive objects. According to measurements, "create + destroy"
of e.g. bittype(32) is at ~6 CPU clock cycles (including function call
overhead), so this is already cheaper than before.