Removed deprecated FIRRTL generation
The FIRRTL generation based on printing ASCII has not been kept up
to date with the CIRCT version, and no longer serves a purpose for
the HLS backend.

The '--circt' argument for jhls has been deprecated but kept to not
break dependent projects, i.e., the hls-test-suite, until they have
removed the use of the argument.