Avoid OutputFormatParsing test failing when MLIR is disabled (#569)
A test was failing when compiling without `--enable-mlir`, due to
`--output-format mlir` being unknown when `ENABLE_MLIR` is not defined.

This issue could have been fixed in a different location, such as not
defining the `Mlir` value in the `OutputFormat` enum at all, but that
would have created some additional situations where `#ifdef`s would have
to be added.

I think it is nice to use as few `#ifdef`s as possible, to avoid
situations like this where tests fail in specific `#ifdef`
configurations. I would not be opposed to including the `--output-format
mlir` option in all builds, but possibly making it invisible in the
help-text.
If `OutputFormat::Mlir` is selected without being enabled in the build,
it eventually triggers a `JLM_UNREACHABLE` that informs you that jlm has
been compiled without the MLIR dialect. I think this is much more
user-friendly than just denying that there ever was an output format
called `mlir`.