Make parser tests use c_str instead of copying strings (#571)
This change simplifies a helper function in
`TestJlcCommandLineParser.cpp` and `TestJlmOptCommandLineParser.cpp` by
using `c_str()`. It is safe to do this as the strings owning the
underlying data live longer than the vector of `const char *` (and the
strings are never modified).

To make this change legal, some `const` had to be added, but only in
places where no mutation is happening.

The `const char * const *` is necessary, as `char **` can not be cast to
`const char **`.
I had to google this to [understand
why](https://isocpp.org/wiki/faq/const-correctness#constptrptr-conversion).