Add missing type check to argument constructor (#574)
The argument constructor did not check whether the type of the provided
input is the same as the handed in type. In the future, the constructor
should be refactored such that we either hand in an input or a simple
type, but not both at the same time. This would avoid this check
altogether.

Closes #555