Add missing type check to result constructors (#575)
The result constructors did not check whether the type of the provided
output is the same as the handed in type. In the future, these
constructors should be refactored such that we either hand in an output
or a simple type, but not both at the same time. This would avoid this
check altogether.

Closes #556