[AndersenAgnostic] Add lazy cycle detection as an optional technique for Andersen (#581)
When propagating along the edge a->b, it is because a has something new.
If b already has everything a has, that might indicate that there is a
path from b to a. Lazy cycle detection looks for such a path using a DFS
from b, and unifies everything that is on such a path b->*->a. It tries
at most once per edge.