lambda: change API for mapping input to region argument
Remove the lambda::cvinput class and the way it is used to map an input
to the corresponding argument inside the lambda region. Replace this with
an API function attached to the node instead.

Rework API around context variable handling: Let API consistently return
"descriptor" of a context variable and formulate everything in these
terms.