[AndersenAgnostic] Add difference propagation (#587)
Adds difference propagation, which is one of the more "invasive"
techniques, as it needs to intercept modifications to the points-to
sets. When iterating over a points-to set, you sometimes only need the
set of pointees that have been added since last visit, while you other
times need all pointees.

For the flags, we get away with not intercepting the setting of them.
Difference propagation can just remember if the flag was set the last
time the work item was visited, which only takes up 2 bits per
PointerObject.