make ClangFormat check HLS, and make it optional to specify CIRCT and MLIR install paths in configure.sh (#591)
This PR started as a fix of formatting that changed with clang-format
18. I also noticed that the constructor had an unnecessary `const`, so I
fixed that.

Looking into why this was not caught by the ClangFormat workflow, I saw
that the `configure.sh` script's `--enable-mlir` "consumed" the
`--enable-hls`, causing hls to not be enabled, which affects the set of
files found by `make format-dry-run`.

In an attempt to make this not happen in the future, I made
`configure.sh` complain and fail if the path you pass to `--enable-mlir`
or `--enable-hls` is not a folder that exists. This forces the user to
set up the folder first, but it should be helpful in preventing this
type of mistake.

Formatting is a bit of a special case, as we want to "enable" hls and
mlir without needing to install either, so I passed in `/usr` as the
path, since that is a folder we know exists.