Add RVSDG tree printer pass (#595)
This pass enables the printing of the RVSDG tree between optimizations.

Intended usage: `jlm-opt --RvsdgTreePrinter --DeadNodeElimination
--RvsdgTreePrinter -s /tmp input.ll`

This will write the files `input-rvsdgTree-0` and `input-rvsdgTree-1` to
`/tmp`, containing the RVSDG tree from before applying dead node
elimination and from after dead node elimination, respectively.

The plan is to extend this pass with support for annotations to the
RVSDG tree. Moreover, the pass can also be generalized to print not just
RVSDG tree, but also the ASCII version of the RVSDG or other debug
output.