Header consistency checking (#619)
make target for header consistency checking

Provide a "check-headers" target that verifies whether all header files
are correctly correctly declared for the build system. This avoids cases
when a header is used, but not declared in any library build rule.

In principle, the build system should always be aware of every single
file
used during build. Otherwise, this can cause problems, e.g. when a
library
is built but the "install" rule misses headers that should go without,
hence publishing a broken package.

The present checking does not verify if the header is "assigned" to the
library
that "uses" it (or some dependent library) -- that is difficult to
express
in "make", but this check should be sufficient for the purposes of
manually
rectifying any oversight.

Fix missing headers detected by check-headers target

Add all headers detected missing by the build target to declarations
in the makefiles.