Move CFG conversion to ascii to cfg class (#631)
The ascii output for the control flow graph is not easily readable. This
is the first PR on improving it. It does the following:

1. Moves the functions for converting a CFG to ascii to the cfg class
2. Adds a unit test for the conversion of three address codes

In order to add a unit test for the CFG conversion, I first need to make
the output more deterministic. This will happen in a follow up PR. This
PR is part of issue #586