Make RegionArgument and RegionResult instantiable (#656)
Make both RegionArgument and RegionResult non-abstract and provide
"Create" functions. This allows slowly phasing out the derived argument
and result classes, replacing them with just RegionArgument and
RegionResult. This allows to remove a lot of boilerplate code as the
number of classes to be implemented and matched shrinks noticably.

Later refactoring should rename the input/output classes and also allow
to clean them up so they do not even need to be virtual anymore.