Add new normalization interface and convert load reductions to it (#683)
This PR is the first in a series of PRs to remove the old node
normalization interface that is intertwined in the RVSDG library. This
PR does the following:

1. Adds a new and simple node normalization interface for optimizing the
operands of a node. This enables to create individual instances of
normalizations and therefore pass them around, and/or configure
optimization by simply providing these instances instead of utilizing
the flags of the old interface.
2. Converts the load reductions to this new interface
3. Utilizes these new reductions in load tests