Distinguish between "function" and "pointer" types
Make the lambda operator return a "function" object (instead of pointer)
and make the call operator require a "function" object (instead of a pointer).
Introduce pointer-to-function and function-to-pointer conversions in places
where references to functions are stored in machine data structures or
passed as machine-representable values across functions.

In pointer analysis, unify functions and pointers to functions as the same
objects. Possibly function objects should be represented as "pointees"
instead, but for now simply blurring the distinction between the two delivers
the desired results.