Function pointer conversion (#705)
Add function / pointer conversion llvm operators. Accept these operators
as top-level nodes in the various transformation passes, and handle them
in instruction conversion.

There are presently no "users" of these operators or anything that puts
them into the graph -- so this is preparatory for making function and
pointer distinction in a later commit.