Make use of debug_string() for arguments and results in dot output (#752)
It was a bit annoying to only see "a1", "r5" in the root RVSDG region,
so this change makes use of "debug_string()" on inputs, outputs, results
and arguments.


![image](https://github.com/user-attachments/assets/1cf6884a-d45a-4294-a5b0-2e2c0c085448)

The `debug_string()` method was not overridden by the subclasses, so I
added that. In general they were not really used anywhere, so adding the
extra info did not break anything existing.

Finally, since I was touching the graph output, I decided to reduce the
margins on argument and result nodes, which were previously very big.