Add general debug output file creation to StatisticsCollector (#746)
Since the `StatisticsCollector` is sent around to the different passes,
I thought it made sense to use it as a general helper for creating debug
output files. The `RvsdgTreePrinterPass` already used the same folder,
but it was manually configured to do so.

The `StatisticsCollectorSettings` now consists of four parts:
 - The demanded statistics.
- The output directory, which is `/tmp/jlm` by default. If the folder
does not exist when a request for an output file is made, the directory
is made (max one dir deep).
- The module name, which becomes the prefix of all files created by the
collector. It is typically the basename of the input file, e.g.,
"test-loops".
- A unique string, which is 6 random alphanumeric chars by default.
Importantly, the same 6 chars are used by all files created by the same
`StatisticsCollector`.

Using this, the `RvsdgTreePrinterPass` now gets the paths for its output
files from the collector. The collector also does the counting when the
pass is called several times.

This change required some added functionality to the `file.hpp`
utilities, with unit tests written.

Lastly, a `StatisticsCollector` is created for `jlm-hls`, and sent in to
the functions running passes.
Actually making `jlm-hls`, `jhls`, and `jlc` take command line
parameters to configure these collectors remains.